ui: dashboard/views: add subscription notice in status row
This adds a short subscription notice to the status row.
It will be loaded together with the normal refresh.

It currently only shows the following states:
* No remotes configured (warning)
* No valid subscriptions (error)
* Valid subscriptions (good)

It reuses the pdm backend subscription check for this.
If there is a message or an error from the api call, show that as
tooltip.

A click on the notice routes to the subscriptions panel of PDM, where
more info about the statistics is outlined.

Downside of this approach is that we might load the subscription info
twice on a fresh login, once for the login popup check, and once for the
dashboard/view if one is open after login.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/20251203095945.1459122-2-d.csapak@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>