ui: default dashboard: don't load/show the subscription notice
we already have the subscription panel there, so showing it again does
not make much sense.

Keep it in the custom views, as there the subscription panel is limited
to the specific view.

To detect if we're in a custom view, use the 'editing_state' property,
which is only set when we can edit the view, and that is not the case
for the default dashboard.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/20251203134050.3215934-1-d.csapak@proxmox.com