ui: view status row: hide subscription notice if valid or no remotes
We do not want to show this here if all is valid, as then the user can
choose themselves if they want to show a subscription info card or
not. E.g., if one creates a view for non-tech-savy managers or the
like, those might be confused by any extra hint, even if positive.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>