ui: subscription info card: clarify that this is the remote status
This is not the resulting status we calculate for the PDM, but rather
purely the status from the remotes, modelled after the PVE Datacenter
overview status.

In the future we might want to rework this a bit into showing the
counts and providing a separate info card for the PDM status here, but
lets not rush this.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>