ui: pve: remove overview: allow line-breaks for long lines
the last two status rows contain longer text than the rest, so it can
easily happen that those get too long for the widget and overflow.

To allow for the text to wrap (without modifying the StatusRow component
itself), wrap the text in a div that allows wrapping and has the right
alignment set.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Reviewed-by: Lukas Wagner <l.wagner@proxmox.com>
Tested-by: Lukas Wagner <l.wagner@proxmox.com>
Link: https://lore.proxmox.com/20251204142451.2992215-2-d.csapak@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>