server: api: pbs: improve error message for datastore content
the code here looks for the response code to determine what to do, but
since the 'extjs' formatter is used, that's always 200, so it's always
attempted to parse the body like a successful api call.

Change the api call to use the 'json' formatter, which correctly sets
the response code directly and not in the body.

This fixes an issue with a very non-informative error message when a pbs
datastore has an error (e.g. no '.chunks' folder) like

 missing field `data` at line 1 column xy

instead of the actual error message from the api call.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>