ui: pve tree: prevent from pushing unnecessary route
In the tree, a route is pushed when the data table selection changes.
When the route changes externally (e.g. the browser back button), we
manually select the correct entry afterwards, which triggers a selection
change, which currently triggers another route push of the same route
that's already active.

This means that on back button click, there will be always a new route
pushed to the same path the user is already on.

Prevent that by comparing the path we want to push with what we're
currently on, and only push it when they differ.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>