cli: admin: add `support-eligibility` subcommand
add cli methods to allow updating and checking the subscription
information of pdm itself. before, the cli only had a cli method that
prints the subscription status for each remote without providing
information on whether the pdm host itself is eligible for support.

the term `support-eligiblity` for the subcommand was chosen over
`subscription`, because:

1. there are no subscriptions for pdm itself, only for the remotes it
is connected to. users may be confused that they can't set a
subscription here like they can with other proxmox products

2. avoid confusion with the `remote subscriptions` subcommand

Signed-off-by: Shannon Sterz <s.sterz@proxmox.com>
Reviewed-by: Lukas Wagner <l.wagner@proxmox.com>
Tested-by: Maximiliano Sandoval <m.sandoval@proxmox.com>
Link: https://lore.proxmox.com/20251215162129.403124-3-s.sterz@proxmox.com