ui: node url list: validate list of endpoints
otherwise user can manually edit these fields. when adding a remote
through a wizzard the api will return an error that the value doesn't
match the validation regex instantly. hence, there it is mostly a ux
issue.

when editing a remote, however, the api doesn't validate the url. so
adding, for a example a protocol (e.g. "https://") to the hostname/ip
field, will lead to connection errors when looking at the remote's
overview.

Signed-off-by: Shannon Sterz <s.sterz@proxmox.com>
Tested-by: Lukas Wagner <l.wagner@proxmox.com>
Reviewed-by: Lukas Wagner <l.wagner@proxmox.com>
Link: https://lore.proxmox.com/20251219123758.151318-2-s.sterz@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>